ParallelFor(Tag)
We have been using a ParallelFor function template that takes a Vector of
Tags in communication functions.  It launches a single GPU kernel for an
irregular iteration space.  This function is useful in many situations.
Therefore, it is moved from a FillBoundary implementation header to its own
header.  And we have added two more versions of this for 1d and 3d loops.
