+++ Src/Base/AMReX_FBI.H
@@ -8,162 +8,10 @@ struct FabCopyTag {
     IntVect offset; // sbox.smallEnd() - dbox.smallEnd()
 };
 
-template <class T>
-struct Array4CopyTag {
-    Array4<T      > dfab;
-    Array4<T const> sfab;
-    Box dbox;
-    Dim3 offset; // sbox.smallEnd() - dbox.smallEnd()
-
-    AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-    Box const& box () const noexcept { return dbox; }
-};
-
-template <class T>
-struct Array4MaskCopyTag {
-    Array4<T      > dfab;
-    Array4<T const> sfab;
-    Array4<int    > mask;
-    Box dbox;
-    Dim3 offset; // sbox.smallEnd() - dbox.smallEnd()
-
-    AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-    Box const& box () const noexcept { return dbox; }
-};
-
 struct VoidCopyTag {
     char const* p;
     Box dbox;
 };
-
-template <class T>
-struct Array4BoxTag {
-    Array4<T> dfab;
-    Box       dbox;
-
-    AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-    Box const& box () const noexcept { return dbox; }
-};
-
-template <class T>
-struct Array4BoxValTag {
-    Array4<T> dfab;
-    Box       dbox;
-    T          val;
-
-    AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-    Box const& box () const noexcept { return dbox; }
-};
-
-#ifdef AMREX_USE_GPU
-namespace detail {
-template <class TagType, class F>
-void
-ParallelFor_doit (Vector<TagType> const& tags, F && f)
-{
-    const int ntags = tags.size();
-    if (ntags == 0) return;
-
-    int ntotwarps = 0;
-    Vector<int> nwarps;
-    nwarps.reserve(ntags+1);
-    for (int i = 0; i < ntags; ++i)
-    {
-        auto& tag = tags[i];
-        nwarps.push_back(ntotwarps);
-        ntotwarps += static_cast<int>(tag.box().numPts()+Gpu::Device::warp_size-1)/Gpu::Device::warp_size;
-    }
-    nwarps.push_back(ntotwarps);
-
-    std::size_t sizeof_tags = ntags*sizeof(TagType);
-    std::size_t offset_nwarps = Arena::align(sizeof_tags);
-    std::size_t sizeof_nwarps = (ntags+1)*sizeof(int);
-    std::size_t total_buf_size = offset_nwarps + sizeof_nwarps;
-
-    char* h_buffer = (char*)The_Pinned_Arena()->alloc(total_buf_size);
-    char* d_buffer = (char*)The_Arena()->alloc(total_buf_size);
-
-    std::memcpy(h_buffer, tags.data(), sizeof_tags);
-    std::memcpy(h_buffer+offset_nwarps, nwarps.data(), sizeof_nwarps);
-    Gpu::htod_memcpy_async(d_buffer, h_buffer, total_buf_size);
-
-    auto d_tags = reinterpret_cast<TagType*>(d_buffer);
-    auto d_nwarps = reinterpret_cast<int*>(d_buffer+offset_nwarps);
-
-    constexpr int nthreads = 256;
-    constexpr int nwarps_per_block = nthreads/Gpu::Device::warp_size;
-    int nblocks = (ntotwarps + nwarps_per_block-1) / nwarps_per_block;
-
-    amrex::launch(nblocks, nthreads, Gpu::gpuStream(),
-#ifdef AMREX_USE_DPCPP
-    [=] AMREX_GPU_DEVICE (sycl::nd_item<1> const& item) noexcept
-    AMREX_REQUIRE_SUBGROUP_SIZE(Gpu::Device::warp_size)
-#else
-    [=] AMREX_GPU_DEVICE () noexcept
-#endif
-    {
-#ifdef AMREX_USE_DPCPP
-        int g_tid = item.get_global_id(0);
-#else
-        int g_tid = blockDim.x*blockIdx.x + threadIdx.x;
-#endif
-        int g_wid = g_tid / Gpu::Device::warp_size;
-        if (g_wid >= ntotwarps) return;
-
-        int tag_id = amrex::bisect(d_nwarps, 0, ntags, g_wid);
-
-        auto tag = d_tags[tag_id];
-        int ncells = tag.box().numPts();
-        int b_wid = g_wid - d_nwarps[tag_id]; // b_wid'th warp on this box
-#ifdef AMREX_USE_DPCPP
-        int lane = item.get_local_id(0) % Gpu::Device::warp_size;
-#else
-        int lane = threadIdx.x % Gpu::Device::warp_size;
-#endif
-        int icell = b_wid*Gpu::Device::warp_size + lane;
-
-        const auto len = amrex::length(tag.box());
-        const auto lo  = amrex::lbound(tag.box());
-        int k =  icell /   (len.x*len.y);
-        int j = (icell - k*(len.x*len.y)) /   len.x;
-        int i = (icell - k*(len.x*len.y)) - j*len.x;
-        i += lo.x;
-        j += lo.y;
-        k += lo.z;
-
-#ifdef AMREX_USE_DPCPP
-        f(item, icell, ncells, i, j, k, tag);
-#else
-        f(icell, ncells, i, j, k, tag);
-#endif
-    });
-
-    Gpu::synchronize();
-    The_Pinned_Arena()->free(h_buffer);
-    The_Arena()->free(d_buffer);
-}
-}
-
-template <class TagType, class F>
-void
-ParallelFor (Vector<TagType> const& tags, int ncomp, F && f)
-{
-    detail::ParallelFor_doit(tags,
-        [=] AMREX_GPU_DEVICE (
-#ifdef AMREX_USE_DPCPP
-            sycl::nd_item<1> const& /*item*/,
-#endif
-            int icell, int ncells, int i, int j, int k, TagType const& tag) noexcept
-        {
-            if (icell < ncells) {
-                for (int n = 0; n < ncomp; ++n) {
-                    f(i,j,k,n,tag);
-                }
-            }
-        });
-}
-
-#endif
 
 namespace detail {
